# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "BossFi区块链招聘论坛后端API",
        "title": "BossFi Backend API",
        "termsOfService": "https://bossfi.io/terms",
        "contact": {
            "name": "BossFi Team",
            "url": "https://bossfi.io",
            "email": "support@bossfi.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/stakes/distribute": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "分发质押奖励给所有符合条件的用户（管理员功能）",
                "tags": [
                    "管理员"
                ],
                "summary": "分发奖励",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.DistributeRewardsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/admin/stakes/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取系统质押统计信息（管理员功能）",
                "tags": [
                    "管理员"
                ],
                "summary": "获取质押统计信息",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/repository.StakeStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stakes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新的质押记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "质押"
                ],
                "summary": "创建质押",
                "responses": {
                    "201": {
                        "description": "成功",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/stakes/rewards/claim": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "领取用户的质押奖励",
                "tags": [
                    "质押"
                ],
                "summary": "领取奖励",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.ClaimRewardsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stakes/user/{user_id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定用户的质押列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "质押"
                ],
                "summary": "获取用户质押列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.ListStakesResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stakes/user/{user_id}/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取指定用户的质押统计信息",
                "tags": [
                    "质押"
                ],
                "summary": "获取用户质押统计信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/repository.UserStakeStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stakes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取质押详细信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "质押"
                ],
                "summary": "获取质押详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "质押ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/stake.Stake"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "质押不存在",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stakes/{id}/complete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "完成解质押指定的质押记录",
                "tags": [
                    "质押"
                ],
                "summary": "完成解质押",
                "parameters": [
                    {
                        "type": "string",
                        "description": "质押ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/stake.Stake"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "质押不存在",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/stakes/{id}/unstake": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "请求解质押指定的质押记录",
                "tags": [
                    "质押"
                ],
                "summary": "请求解质押",
                "parameters": [
                    {
                        "type": "string",
                        "description": "质押ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/stake.Stake"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "质押不存在",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Response": {
            "description": "API统一响应格式",
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应码，0表示成功",
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "响应消息",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "repository.StakeStats": {
            "type": "object",
            "properties": {
                "active_stakes": {
                    "type": "integer"
                },
                "average_amount": {
                    "type": "number"
                },
                "completed_stakes": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_rewards": {
                    "type": "number"
                },
                "total_stakes": {
                    "type": "integer"
                },
                "unstaking_stakes": {
                    "type": "integer"
                }
            }
        },
        "repository.UserStakeStats": {
            "type": "object",
            "properties": {
                "active_amount": {
                    "type": "number"
                },
                "active_stakes": {
                    "type": "integer"
                },
                "completed_stakes": {
                    "type": "integer"
                },
                "pending_rewards": {
                    "type": "number"
                },
                "total_amount": {
                    "type": "number"
                },
                "total_rewards": {
                    "type": "number"
                },
                "total_stakes": {
                    "type": "integer"
                },
                "unstaking_stakes": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "service.ClaimRewardsResponse": {
            "type": "object",
            "properties": {
                "claimed_amount": {
                    "type": "number"
                },
                "new_balance": {
                    "type": "number"
                }
            }
        },
...
        "service.DistributeRewardsResponse": {
            "type": "object",
            "properties": {
                "failed_count": {
                    "type": "integer"
                },
                "processed_count": {
                    "type": "integer"
                },
                "total_reward": {
                    "type": "number"
                }
            }
        },
        "service.ListStakesResponse": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "stakes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/stake.Stake"
                    }
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "stake.Stake": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_reward_at": {
                    "type": "string"
                },
                "reward_earned": {
                    "type": "number"
                },
                "staked_at": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/stake.StakeStatus"
                },
                "unstake_request_at": {
                    "type": "string"
                },
                "unstaked_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "stake.StakeStatus": {
            "type": "string",
            "enum": [
                "active",
                "unstaking",
                "completed"
            ],
            "x-enum-comments": {
                "StakeStatusActive": "活跃",
                "StakeStatusCompleted": "已完成",
                "StakeStatusUnstaking": "解质押中"
            },
            "x-enum-varnames": [
                "StakeStatusActive",
                "StakeStatusUnstaking",
                "StakeStatusCompleted"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}